GPII-2732: Integrate  for trigger conditions
Single rule engine functioning
