+++ src/main/surveys/surveyServer.js
@@ -12,11 +12,13 @@ wss.on("connection", function connection(ws) {
 
         var triggerFixture = {
             surveyTrigger: {
-                conditions: [
-                    {
-                        "minutesSinceKeyIn": 3
-                    }
-                ],
+                conditions: {
+                    all: [{
+                        fact: "keyedInBefore",
+                        operator: "greaterThanInclusive",
+                        value: 1000 * 5 // 5 secs
+                    }]
+                },
                 id: "id",
                 urlTriggerHandler: "URL of the survey server to handle the surveyTriggerEvent"
             }
